chore: Further extraction of classes from common modules
