+++ ra2ce/analyses/configuration/readers/analysis_config_reader_base.py
@@ -25,7 +25,7 @@ from pathlib import Path
 from ra2ce.common.configuration.ini_configuration_reader_protocol import (
     IniConfigurationReaderProtocol,
 )
-from ra2ce.configuration.validators.ini_config_validator_base import (
+from ra2ce.analyses.configuration.analysis_ini_config_data_validator_base import (
     DirectAnalysisNameList,
     IndirectAnalysisNameList,
 )
