+++ ra2ce/common/configuration/config_protocol.py
@@ -25,11 +25,11 @@ from __future__ import annotations
 from pathlib import Path
 from typing import Any, List, Optional, Protocol, runtime_checkable
 
-from ra2ce.configuration.ini_config_protocol import IniConfigDataProtocol
+from ra2ce.common.configuration.config_data_protocol import IniConfigDataProtocol
 
 
 @runtime_checkable
-class ConfigProtocol(Protocol):  # pragma: no cover
+class ConfigWrapperProtocol(Protocol):  # pragma: no cover
     ini_file: Path
     root_dir: Path
     config_data: Optional[IniConfigDataProtocol] = None
@@ -38,7 +38,7 @@ class ConfigProtocol(Protocol):  # pragma: no cover
     @classmethod
     def from_data(
         cls, ini_file: Path, config_data: IniConfigDataProtocol
-    ) -> ConfigProtocol:
+    ) -> ConfigWrapperProtocol:
         """
         Initializes a `ConfigProtocol` with the given parameters.
