+++ ra2ce/graph/network_config_data/network_config_data_validator.py
@@ -27,7 +27,9 @@ from ra2ce.graph.network_config_data.network_config_data import (
 )
 from ra2ce.common.validation.ra2ce_validator_protocol import Ra2ceIoValidator
 from ra2ce.common.validation.validation_report import ValidationReport
-from ra2ce.configuration.validators.ini_config_validator_base import _expected_values
+from ra2ce.analyses.configuration.analysis_ini_config_data_validator_base import (
+    _expected_values,
+)
 
 
 class NetworkConfigDataValidator(Ra2ceIoValidator):
