+++ tests/analyses/configuration/test_analysis_ini_config_data_validator_base.py
@@ -3,8 +3,8 @@ from typing import List
 
 import pytest
 
-from ra2ce.configuration.ini_config_protocol import IniConfigDataProtocol
-from ra2ce.configuration.validators.ini_config_validator_base import (
+from ra2ce.common.configuration.config_data_protocol import IniConfigDataProtocol
+from ra2ce.analyses.configuration.analysis_ini_config_data_validator_base import (
     IniConfigValidatorBase,
 )
 from ra2ce.common.validation.validation_report import ValidationReport
